完成第二次代码重构，区分管理员、会员和浏览者的功能

+++ member/articles.py
@@ -1,10 +1,11 @@
 from flask import request, redirect, url_for, render_template, session
 from libs import db
 from models import Article
-from .admin_app import admin_app
+from .member_app import member_app
+import json
 
 
-@admin_app.route("/article/post", methods=['get', 'post'])
+@member_app.route("/article/post", methods=['get', 'post'])
 def article_post():
     if request.method == "POST":
         cate_id = request.form['cate']
@@ -20,14 +21,13 @@ def article_post():
         )
         db.session.add(article)
         db.session.commit()
-        import json
         message = {"message": "文章发布成功"}
         return json.dumps(message)
-    return render_template("admin/article/article_post.html")
+    return render_template("member/article/article_post.html")
 
 
-@admin_app.route("/article/list/<int:page>", methods=['get', 'post'])
-@admin_app.route("/article/list", defaults={"page": 1}, methods=['get', 'post'])
+@member_app.route("/article/list/<int:page>", methods=['get', 'post'])
+@member_app.route("/article/list", defaults={"page": 1}, methods=['get', 'post'])
 def article_list(page):
     if request.method == 'POST':
         q = request.form['q']
@@ -40,37 +40,44 @@ def article_list(page):
             order = Article.id.asc()
         else:
             order = Article.id.desc()
-        res = Article.query.filter(condition).order_by(order).paginate(page, 10)
+        res = Article.query.filter(Article.author == session['user']).filter(condition).order_by(order).paginate(page,
+                                                                                                                 10)
     else:
-        res = Article.query.paginate(page, 10)
+        res = Article.query.filter(Article.author == session['user']).paginate(page, 10)
     # 无论搜索还是默认查看，都是翻页处理
     articles = res.items
     pageList = res.iter_pages()
     total = res.total
     pages = res.pages
-    return render_template("admin/article/article_list.html", articles=articles, pageList=pageList, total=total,
+    return render_template("member/article/article_list.html", articles=articles, pageList=pageList, total=total,
                            pages=pages)
 
 
 # 根据文章id删除文章
-@admin_app.route('/article/delete/<int:article_id>')
+@member_app.route('/article/delete/<int:article_id>')
 def article_delete(article_id):
     article = Article.query.get(article_id)
-    db.session.delete(article)
-    db.session.commit()
+    # 只能删除自己的文章
+    if article.author == session['user']:
+        db.session.delete(article)
+        db.session.commit()
     return redirect(url_for('.article_list'))
 
 
 # 文章修改
-@admin_app.route('/article/edit/<int:article_id>', methods=['get', 'post'])
+@member_app.route('/article/edit/<int:article_id>', methods=['get', 'post'])
 def article_edit(article_id):
     article = Article.query.get(article_id)
-    if request.method == 'POST':
-        article.cate_id = request.form['cate']
-        article.title = request.form['title']
-        article.intro = request.form['intro']
-        article.content = request.form['content']
-        db.session.add(article)
-        db.session.commit()
+    if not article:
         return redirect(url_for('.article_list'))
-    return render_template('admin/article/article_edit.html', article=article)
+    if request.method == 'POST':
+        # 只能修改自己的文章
+        if article.author == session['user']:
+            article.cate_id = request.form['cate']
+            article.title = request.form['title']
+            article.intro = request.form['intro']
+            article.content = request.form['content']
+            db.session.add(article)
+            db.session.commit()
+            return redirect(url_for('.article_list'))
+    return render_template('member/article/article_edit.html', article=article)
